Move common proposal and vote implementation so that cw3-flex-multisig would import it from cw3-fixed-multisig
