+++ contracts/cw3-fixed-multisig/Cargo.toml
@@ -21,7 +21,6 @@ library = []
 utils = { path = "../../packages/utils", version = "0.10.3" }
 cw2 = { path = "../../packages/cw2", version = "0.10.3" }
 cw3 = { path = "../../packages/cw3", version = "0.10.3" }
-cw3-flex-multisig = { path = "../cw3-flex-multisig", version = "0.10.3" }
 cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3" }
 cosmwasm-std = { version = "1.0.0-beta3" }
 schemars = "0.8.1"
